\name{r2time-package}
\alias{r2time-package}
\alias{r2time}
\docType{package}
\title{
Analysis of Large opentsdb time-series datasets in HBase
}
\description{
Analysis of Large opentsdb time-series datasets in HBase
}
\details{
\tabular{ll}{
Package: \tab r2time\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-07-16\cr

}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Bikash Agrawal

}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}

